// src/internal/types/messaging/messaging-types.ts
export type MailBoxCapabilitiesJson = {
    /**
     * Whether the voicemail server can return the list of messages.
     */
    listMessages?: boolean;

    /**
     * Whether Voice messages can be downloaded.
     */
    getMessages?: boolean;

    /**
     * Whether Recorded messages can be downloaded.
     */
    getRecord?: boolean;

    /**
     * Whether the voicemail server can play voice messages.
     */
    play?: boolean;

    /**
     * Whether a played voice message can be paused and resumes from the position it has been paused.
     */
    pause?: boolean;

    /**
     * Whether the media session can be terminate.
     */
    hangup?: boolean;

    /**
     * Whether the voicemail server can record voice messages.
     */
    record?: boolean;

    /**
     * Wheter a voice message can be resumed to the position it has been paused.
     */
    resume?: boolean;

    /**
     * Whether the current recording can be cancelled.
     */
    cancel?: boolean;

    /**
     * Wheter the voicemail server can forward voice messages.
     */
    forward?: boolean;

    /**
     * Whether the voicemail server can call back the originator of the voice message.
     */
    callback?: boolean;

    /**
     * Whether a voice message or a record can be sent to recipients.
     */
    send?: boolean;

    /**
     * Whether the voicemail server can send events in case of message deposit / removal.
     */
    events?: boolean;
};

/**
 * MailBox represents a mail box in a voice mail system. A voice mail is assigned to a user.
 */
/** @internal */
export type MailBoxJson = {
    /**
     * The mail box identifier.
     */
    id?: string;

    /**
     * The mail box name.
     */
    name?: string;

    /**
     * The mail box capabilities.
     */
    capabilities?: MailBoxCapabilitiesJson;
};